+++ CHANGELOG.md
@@ -8,3 +8,6 @@ All notable changes to the "Snipped" extension will be documented in this file.
 - 0.19.0 Adding Touch Bar shortcut
 - 0.22.0 Adding SVG as default format & AutoCopy option
 - 1.0.2 Removing full path & updating separator + render lib
+- 1.3.0 Added scale settings, remove title setting, update security libs and update core
+- Added donation address for crypto BSC
+- ### Thank you for 12k downloads!!!